Reverted all_fields and pagination test
